+++ Picroup-iOS/Routers/Router.swift
@@ -50,7 +50,7 @@ final class Router {
         
         _ = store.reputationsRoute().distinctUntilChanged { $0.version ?? "" }.skip(1)
             .drive(Binder(self) { (me, _) in
-                let vc = RouterService.Main.reputationsViewController()
+                let vc = RouterService.Me.reputationsViewController()
                 me.currentNavigationController?.pushViewController(vc, animated: true)
             })
         
@@ -76,11 +76,11 @@ final class Router {
                 let vc: UIViewController
                 switch (isCurrentUser, userRoute.userId) {
                 case (true, _):
-                    vc = RouterService.Main.meViewController()
+                    vc = RouterService.Me.meViewController()
                     vc.hidesBottomBarWhenPushed = true
                     me.currentNavigationController?.pushViewController(vc, animated: true)
                 case (false, let userId?):
-                    vc = RouterService.Main.userViewController(dependency: userId)
+                    vc = RouterService.Me.userViewController(dependency: userId)
                     me.currentNavigationController?.pushViewController(vc, animated: true)
                 default:
                     break
@@ -89,27 +89,27 @@ final class Router {
         
         _ = store.updateUserRoute().distinctUntilChanged { $0.version ?? "" }.skip(1)
             .drive(Binder(self) { (me, _) in
-                let vc = RouterService.Main.updateUserViewController()
+                let vc = RouterService.Me.updateUserViewController()
                 me.currentNavigationController?.pushViewController(vc, animated: true)
             })
         
         _ = store.userFollowingsRoute().distinctUntilChanged { $0.version ?? "" }.skip(1)
             .map { $0.userId }.unwrap()
             .drive(Binder(self) { (me, userId) in
-                let vc = RouterService.Main.followingsViewController(dependency: userId)
+                let vc = RouterService.Me.followingsViewController(dependency: userId)
                 me.currentNavigationController?.pushViewController(vc, animated: true)
             })
         
         _ = store.userFollowersRoute().distinctUntilChanged { $0.version ?? "" }.skip(1)
             .map { $0.userId }.unwrap()
             .drive(Binder(self) { (me, userId) in
-                let vc = RouterService.Main.followersViewController(dependency: userId)
+                let vc = RouterService.Me.followersViewController(dependency: userId)
                 me.currentNavigationController?.pushViewController(vc, animated: true)
             })
         
         _ = store.searchUserRoute().distinctUntilChanged { $0.version ?? "" }.skip(1)
             .drive(Binder(self) { (me, _) in
-                let vc = RouterService.Main.searchUserViewController()
+                let vc = RouterService.Me.searchUserViewController()
                 me.currentNavigationController?.pushViewController(vc, animated: true)
             })
         
@@ -151,8 +151,8 @@ final class Router {
                     result.viewControllers = [
                         RouterService.Main.homeMenuViewController(),
                         RouterService.Main.rankViewController(),
-                        RouterService.Main.notificationsViewController(),
-                        RouterService.Main.meNavigationViewController(),
+                        RouterService.Me.notificationsViewController(),
+                        RouterService.Me.meNavigationViewController(),
                     ]
                     return result
                 }()
@@ -166,8 +166,8 @@ final class Router {
                 mainTabBarController.viewControllers = [
                     RouterService.Main.homeMenuViewController(),
                     RouterService.Main.rankViewController(),
-                    RouterService.Main.notificationsViewController(),
-                    RouterService.Main.meNavigationViewController(),
+                    RouterService.Me.notificationsViewController(),
+                    RouterService.Me.meNavigationViewController(),
                 ]
                 
             }
