+++ Picroup-iOS/Routers/RouterService+Main.swift
@@ -25,8 +25,8 @@ extension RouterService.Main {
         mvc.viewControllers = [
             homeMenuViewController(),
             rankViewController(),
-            notificationsViewController(),
-            meNavigationViewController(),
+            RouterService.Me.notificationsViewController(),
+            RouterService.Me.meNavigationViewController(),
         ]
         mvc.tabBar.isTranslucent = false
 //        mvc.selectedIndex = 1
@@ -55,56 +55,4 @@ extension RouterService.Main {
         return bnvc
     }
     
-    static func notificationsViewController() -> UIViewController {
-        let nvc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "NotificationsViewController") as! NotificationsViewController
-        let bnvc = BaseNavigationController(rootViewController: nvc)
-        bnvc.tabBarItem = UITabBarItem(title: "通知", image: UIImage(named: "ic_notifications"), selectedImage: nil)
-        return bnvc
-    }
-    
-    static func meNavigationViewController() -> UIViewController {
-        let mevc = meViewController()
-        let bnvc = BaseNavigationController(rootViewController: mevc)
-        bnvc.isNavigationBarHidden = true
-        bnvc.tabBarItem = UITabBarItem(title: "我", image: UIImage(named: "ic_person"), selectedImage: nil)
-        return bnvc
-    }
-    
-    static func meViewController() -> MeViewController {
-        let mevc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "MeViewController") as! MeViewController
-        return mevc
-    }
-    
-    static func userViewController(dependency: UserViewController.Dependency) -> UserViewController {
-        let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "UserViewController") as! UserViewController
-        vc.dependency = dependency
-        return vc
-    }
-    
-    static func updateUserViewController() -> UpdateUserViewController {
-        let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "UpdateUserViewController") as! UpdateUserViewController
-        return vc
-    }
-    
-    static func reputationsViewController() -> ReputationsViewController {
-        let rvc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "ReputationsViewController") as! ReputationsViewController
-        return rvc
-    }
-    
-    static func followingsViewController(dependency: FollowingsViewController.Dependency) -> FollowingsViewController {
-        let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "FollowingsViewController") as! FollowingsViewController
-        vc.dependency = dependency
-        return vc
-    }
-    
-    static func searchUserViewController() -> SearchUserViewController {
-        let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "SearchUserViewController") as! SearchUserViewController
-        return vc
-    }
-    
-    static func followersViewController(dependency: FollowingsViewController.Dependency) -> FollowersViewController {
-        let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "FollowersViewController") as! FollowersViewController
-        vc.dependency = dependency
-        return vc
-    }
 }
